Cleaned-up extraction scripts putting in common shared codes
