+++ scripts/extractions/extract_common.py
@@ -1,0 +1,191 @@
+"""Common functions used by extraction scripts."""
+
+import json
+import logging
+import os
+from datetime import datetime
+from importlib.metadata import version
+from pathlib import Path
+from tempfile import NamedTemporaryFile
+from typing import List
+
+import geojson
+import geopandas as gpd
+import pandas as pd
+import pystac
+import requests
+from openeo_gfmap.utils.netcdf import update_nc_attributes
+from shapely import Point
+
+# Logger used for the pipeline
+pipeline_log = logging.getLogger("pipeline_sar")
+
+pipeline_log.setLevel(level=logging.INFO)
+
+stream_handler = logging.StreamHandler()
+pipeline_log.addHandler(stream_handler)
+
+formatter = logging.Formatter("%(asctime)s|%(name)s|%(levelname)s:  %(message)s")
+stream_handler.setFormatter(formatter)
+
+
+# Exclude the other loggers from other libraries
+class ManagerLoggerFilter(logging.Filter):
+    """Filter to only accept the OpenEO-GFMAP manager logs."""
+
+    def filter(self, record):
+        return record.name in [pipeline_log.name]
+
+
+stream_handler.addFilter(ManagerLoggerFilter())
+
+
+def post_job_action(
+    job_items: List[pystac.Item], row: pd.Series, parameters: dict = {}
+) -> list:
+    """From the job items, extract the metadata and save it in a netcdf file."""
+    base_gpd = gpd.GeoDataFrame.from_features(json.loads(row.geometry)).set_crs(
+        epsg=4326
+    )
+    assert len(base_gpd[base_gpd.extract == 2]) == len(
+        job_items
+    ), "The number of result paths should be the same as the number of geometries"
+
+    extracted_gpd = base_gpd[base_gpd.extract == 2].reset_index(drop=True)
+    # In this case we want to burn the metadata in a new file in the same folder as the S2 product
+    for idx, item in enumerate(job_items):
+        if "sample_id" in extracted_gpd.columns:
+            sample_id = extracted_gpd.iloc[idx].sample_id
+        else:
+            sample_id = extracted_gpd.iloc[idx].sampleID
+
+        ref_id = extracted_gpd.iloc[idx].ref_id
+        valid_time = extracted_gpd.iloc[idx].valid_time
+        h3_l3_cell = extracted_gpd.iloc[idx].h3_l3_cell
+        s2_tile = row.s2_tile
+
+        item_asset_path = Path(list(item.assets.values())[0].href)
+
+        # Add some metadata to the result_df netcdf file
+        new_attributes = {
+            "start_date": row.start_date,
+            "end_date": row.end_date,
+            "valid_time": valid_time,
+            "GFMAP_version": version("openeo_gfmap"),
+            "creation_date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+            "description": parameters.get(
+                "description", "Sentinel1 GRD raw observations, unprocessed."
+            ),
+            "title": parameters.get("title", "Sentinel-1 GRD"),
+            "sample_id": sample_id,
+            "ref_id": ref_id,
+            "spatial_resolution": parameters.get("spatial_resolution", "20m"),
+            "s2_tile": s2_tile,
+            "h3_l3_cell": h3_l3_cell,
+        }
+
+        # Saves the new attributes in the netcdf file
+        update_nc_attributes(item_asset_path, new_attributes)
+
+    return job_items
+
+
+def generate_output_path(
+    root_folder: Path, geometry_index: int, row: pd.Series, s2_grid: gpd.GeoDataFrame
+):
+    """Generate the output path for the extracted data, from a base path and
+    the row information.
+    """
+    features = geojson.loads(row.geometry)
+    sample_id = features[geometry_index].properties.get("sample_id", None)
+    if sample_id is None:
+        sample_id = features[geometry_index].properties["sampleID"]
+    ref_id = features[geometry_index].properties["ref_id"]
+
+    if "orbit_state" in row:
+        orbit_state = f"_{row.orbit_state}"
+    else:
+        orbit_state = ""
+
+    s2_tile_id = row.s2_tile
+    h3_l3_cell = row.h3_l3_cell
+    epsg = s2_grid[s2_grid.tile == s2_tile_id].iloc[0].epsg
+
+    subfolder = root_folder / ref_id / h3_l3_cell / sample_id
+    return (
+        subfolder
+        / f"{row.out_prefix}{orbit_state}_{sample_id}_{epsg}_{row.start_date}_{row.end_date}{row.out_extension}"
+    )
+
+
+def buffer_geometry(
+    geometries: geojson.FeatureCollection, distance_m: int = 320
+) -> gpd.GeoDataFrame:
+    """For each geometry of the colleciton, perform a square buffer of 320
+    meters on the centroid and return the GeoDataFrame. Before buffering,
+    the centroid is clipped to the closest 20m multiplier in order to stay
+    aligned with the Sentinel-1 pixel grid.
+    """
+    gdf = gpd.GeoDataFrame.from_features(geometries).set_crs(epsg=4326)
+    utm = gdf.estimate_utm_crs()
+    gdf = gdf.to_crs(utm)
+
+    # Perform the buffering operation
+    gdf["geometry"] = gdf.centroid.apply(
+        lambda point: Point(round(point.x / 20.0) * 20.0, round(point.y / 20.0) * 20.0)
+    ).buffer(
+        distance=distance_m, cap_style=3
+    )  # Square buffer
+
+    return gdf
+
+
+def filter_extract_true(geometries: geojson.FeatureCollection) -> gpd.GeoDataFrame:
+    """Remove all the geometries from the Feature Collection that have the property field `extract` set to `False`"""
+    return geojson.FeatureCollection(
+        [f for f in geometries.features if f.properties.get("extract", 0) == 1]
+    )
+
+
+def upload_geoparquet_artifactory(gdf: gpd.GeoDataFrame, name: str) -> str:
+    """Upload the given GeoDataFrame to artifactory and return the URL of the
+    uploaded file. Necessary as a workaround for Polygon sampling in OpenEO
+    using custom CRS.
+    """
+    # Save the dataframe as geoparquet to upload it to artifactory
+    temporary_file = NamedTemporaryFile()
+    gdf.to_parquet(temporary_file.name)
+
+    artifactory_username = os.getenv("ARTIFACTORY_USERNAME")
+    artifactory_password = os.getenv("ARTIFACTORY_PASSWORD")
+
+    headers = {"Content-Type": "application/octet-stream"}
+
+    upload_url = f"https://artifactory.vgt.vito.be/artifactory/auxdata-public/gfmap-temp/openeogfmap_dataframe_{name}.parquet"
+
+    with open(temporary_file.name, "rb") as f:
+        response = requests.put(
+            upload_url,
+            headers=headers,
+            data=f,
+            auth=(artifactory_username, artifactory_password),
+            timeout=180,
+        )
+
+    assert (
+        response.status_code == 201
+    ), f"Error uploading the dataframe to artifactory: {response.text}"
+
+    return upload_url
+
+
+def get_job_nb_polygons(row: pd.Series) -> int:
+    """Get the number of polygons in the geometry."""
+    return len(
+        list(
+            filter(
+                lambda feat: feat.properties.get("extract"),
+                geojson.loads(row.geometry)["features"],
+            )
+        )
+    )
